requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
supabase>=2.27.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets>=15.0.0
